First iter3 tests & fix iter1/iter2 tests
Because we update `make_camlbrick`, we need to update old tests.
For example it's no longer "ball" but "balls" since there are more that
a single ball on the game (thanks to brick bonus).
Balls position are now reference to vec2 as before we couldn't modify
there position.

Also we add "speed" linked to speed cursor at the right part of the UI
to set ball speed.

Signed-off-by: Max Charrier <m@lilmaax.fr>